Move getUserEventAwardCount from user to site-award
