+++ app_legacy/Helpers/database/site-award.php
@@ -210,3 +210,26 @@ function getRecentMasteryData(string $date, string $friendsOf = null, int $offse
 
     return $retVal;
 }
+
+/**
+ * Gets the number of event awards a user has earned
+ */
+function getUserEventAwardCount(string $user): int
+{
+    $bindings = [
+        'user' => $user,
+        'type' => AwardType::Mastery,
+        'event' => 101,
+    ];
+
+    $query = "SELECT COUNT(DISTINCT AwardData) AS TotalAwards 
+              FROM SiteAwards sa
+              INNER JOIN GameData gd ON gd.ID = sa.AwardData
+              WHERE User = :user              
+              AND AwardType = :type
+              AND gd.ConsoleID = :event";
+
+    $dataOut = legacyDbFetch($query, $bindings);
+
+    return $dataOut['TotalAwards'];
+}
